+++ src/rocketgram/__init__.py
@@ -3,7 +3,6 @@
 # Rocketgram is released under the MIT License (see LICENSE).
 
 
-from . import tools
 from .api import *
 from .bot import Bot
 from .connectors import *
@@ -14,6 +13,7 @@ from .json_adapters import *
 from .keyboards import *
 from .middlewares import *
 from .routers import *
+from . import tools
 from .version import version
 
 API_URL = Connector.API_URL